refactor bookingRequest into service and improve error handling + front-end refactor (#114)
Also, an overhaul of the front-end booking request code. I started by trying to simplify components here but ended quite similar to what we had before... 🙈 I hope you don't mind @wigsnes ❤️ 

I just wanted to mention that the save button has been removed. 
Instead, clicking a period in the calendar books it right away. And it can be toggled back and forth as much as you want. Relying on the state in the back-end instead of local.

A loading state has been added to the button as well when toggling its state.